Split form into separate pages - Reduce the amount of form fields on each page to make it more manageable - Store submitted data in session
