+++ app.js
@@ -4,10 +4,11 @@ var favicon = require('serve-favicon');
 var logger = require('morgan');
 var cookieParser = require('cookie-parser');
 var bodyParser = require('body-parser');
+var session = require('express-session');
 var sassMiddleware = require('node-sass-middleware');
 var mojularSassPaths = require('mojular/sass-paths');
 
-var routes = require('./routes/index');
+var index = require('./routes/index');
 var enquiryForm = require('./routes/enquiry-form');
 
 var app = express();
@@ -30,6 +31,11 @@ app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
+app.use(session({
+  secret: process.env.SECRET,
+  resave: false,
+  saveUninitialized: true
+}));
 app.use(sassMiddleware({
   src: path.join(__dirname, 'public', 'sass'),
   dest: path.join(__dirname, 'public', 'stylesheets'),
@@ -39,7 +45,7 @@ app.use(sassMiddleware({
 }));
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', routes);
+app.use('/', index);
 app.use('/enquire', enquiryForm);
 
 // catch 404 and forward to error handler
