+++ lib/fieldIterator.js
@@ -8,6 +8,10 @@ module.exports = function (name, field, options) {
     classes: ['error-message']
   }, field.error);
 
+  if (field.widget.type === 'hidden') {
+    return field.widget.toHTML(name);
+  }
+
   if (field.widget.type === 'multipleCheckbox' || field.widget.type === 'multipleRadio') {
     var fieldsetAttrs = { classes: [] };
     if (opt.fieldsetClasses) {
