--- lib/fields.js
@@ -1,82 +1,0 @@
-var forms = require('forms');
-var fields = forms.fields;
-var validators = forms.validators;
-var widgets = forms.widgets;
-
-var utils = require('./utils');
-var countries = require('../data/countries');
-var industries = require('../data/industries');
-
-module.exports = forms.create({
-  name: fields.string({
-    label: 'Your full name',
-    required: validators.required('is required'),
-  }),
-  'job-title': fields.string({
-    required: true,
-  }),
-  'email-address': fields.email({
-    required: validators.required('is required'),
-  }),
-  'telephone-number': fields.string({
-    required: validators.required('is required'),
-  }),
-  'company-name': fields.string({
-    required: validators.required('is required'),
-  }),
-  'company-address': fields.string({
-    required: validators.required('is required'),
-    widget: widgets.textarea({ rows: 4 }),
-  }),
-  'company-country': fields.array({
-    label: 'Which country is your company based in?',
-    required: validators.required('is required'),
-    widget: widgets.select({
-      classes: ['js-autocomplete']
-    }),
-    choices: utils.mapOptions(countries, { insertBlank: true }),
-  }),
-  'company-industry': fields.string({
-    label: 'Primary operating industry',
-    required: validators.required('is required'),
-    widget: widgets.select({
-      classes: ['js-autocomplete']
-    }),
-    choices: utils.mapOptions(industries, { insertBlank: true })
-  }),
-  'annual-turnover': fields.string({
-    label: 'Annual turnover (optional)',
-  }),
-  'number-of-employees': fields.string({
-    label: 'Number of employees (optional)',
-  }),
-  'enquiry-summary': fields.string({
-    label: 'Provide a brief summary of the assistance you require',
-    required: validators.required('is required'),
-    widget: widgets.textarea({ rows: 6 }),
-  }),
-  'how-you-heard-about-us': fields.array({
-    required: validators.required('is required'),
-    label: 'How did you hear about UK Trade & Investment?',
-    widget: widgets.multipleRadio(),
-    choices: {
-      'Contact with UKTI staff member': 'Contact with UKTI staff member',
-      'GOV.UK website': 'GOV.UK website',
-      'Press article/advert': 'Press article/advert',
-      'Intermediary (for example a trade association, chamber or bank)': 'Intermediary (for example a trade association, chamber or bank)',
-      'Social Media': 'Social Media',
-      'Personal reference': 'Personal reference',
-      'Other': 'Other',
-    },
-  }),
-  'data-protection': fields.array({
-    required: validators.required('is required'),
-    label: 'Would you be happy to be contacted again?',
-    widget: widgets.multipleCheckbox(),
-    choices: {
-      'By UK Trade & Investment': 'By UK Trade & Investment',
-      'By external UK Trade & Investment partners': 'By external UK Trade & Investment partners',
-      'By overseas providers for UK Trade & Investment services': 'By overseas providers for UK Trade & Investment services',
-    },
-  }),
-}, { validatePastFirstError: true });
