+++ lib/mailer.js
@@ -1,4 +1,5 @@
 var q = require('q');
+var S = require('string');
 var nodemailer = require('nodemailer');
 var htmlToText = require('nodemailer-html-to-text').htmlToText;
 var jadeCompiler = require('../lib/jadeCompiler');
@@ -19,7 +20,7 @@ module.exports = {
   sendEnquiry: function (formData) {
     var deferred = q.defer();
 
-    jadeCompiler.compile('emails/inward-enquiry', { fields: formData }, function (error, html) {
+    jadeCompiler.compile('emails/inward-enquiry', { data: formData, S: S }, function (error, html) {
       if (error) {
         deferred.reject(error);
       } else {
