+++ package.json
@@ -11,6 +11,7 @@
     "cookie-parser": "~1.3.5",
     "debug": "~2.2.0",
     "express": "~4.12.4",
+    "express-session": "^1.12.1",
     "forms": "^1.1.4",
     "jade": "~1.9.2",
     "lodash": "^3.10.1",
