+++ views/emails/inward-enquiry.jade
@@ -1,7 +1,9 @@
 p A new enquiry has been received.
-table(width='100%')
-  tbody
-    each value, field in fields
-      tr
-        th(align='left' width='20%')= field
-        td= value
+each fields, section in data
+  h2 #{S(section).humanize().s}
+  table(width='100%')
+    tbody
+      each value, name in fields
+        tr
+          th(align='left' width='180') #{S(name).humanize().s}
+          td= value
