+++ views/enquiry-form.jade
@@ -1,7 +1,7 @@
 extends layout
 
 block content
-  form(action='/enquire' method='post' novalidate)
+  form(action='' method='post' novalidate)
     if errors
       .error-summary(role="group" aria-labelledby="error-summary" tabindex="-1")
         h1#error-summary.heading-medium.error-summary-heading There was a problem submitting this form
@@ -10,7 +10,10 @@ block content
           for field in errors
             li
               a(href='enquire#id_#{field.name}') #{S(field.name).humanize().s} #{field.error}
-    h1 Investment enquiry
+    h1= heading
     != form
     .form-group
-      button.button(type='submit') Submit enquiry
+      p
+        button.button(type='submit')= lastStep ? 'Submit enquiry' : 'Continue'
+      if prevStep
+        a(href='/enquire/#{prevStep}') Back
