+++ views/layout.jade
@@ -4,7 +4,7 @@ block head
   base(href='/')
 
 block prepend page_title
-  = title + ' | '
+  = 'Contact UK Trade & Investment | '
 
 block stylesheets
   link(rel='stylesheet', href='/stylesheets/style.css')
